Correct issues with onBeforePrint
